requirements.txt
streamlit
python-docx
python-pptx
reportlab
pandas
openpyxl
google-generativeai
